updated_backend1/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-SocketIO
SQLAlchemy
psycopg2-binary
boto3
botocore
PyJWT
Werkzeug
python-dotenv
marshmallow
Flask-Marshmallow
marshmallow-sqlalchemy
email-validator
phonenumbers
Pillow
gunicorn
alembic
pytz
qrcode
redis
celery
pytest
pytest-flask
black
flake8
mypy
# Missing these critical packages:
firebase-admin
google-cloud-firestore